# [DNM] implement Quo using Newton-Raphson