In attempting to work on #22, I implemented Newton-Raphson. I got around the exactness problem the same way cuberoot does: multiply the result by the input and see if it's an exact match. All of the tests pass with this code.

However I was not able to observe any performance increase. Instead, I saw a huge performance decrease:

```
name                old time/op  new time/op     delta
GDA/divide-8        1.49ms ± 1%     9.21ms ± 1%    +517.48%  (p=0.008 n=5+5)
GDA/divideint-8      373µs ± 0%      393µs ± 1%      +5.45%  (p=0.008 n=5+5)
GDA/exp-8            262ms ± 2%      339ms ± 3%     +29.02%  (p=0.008 n=5+5)
GDA/ln-8             162ms ± 2%      291ms ± 2%     +79.24%  (p=0.008 n=5+5)
GDA/log10-8          206ms ± 3%      362ms ± 1%     +75.51%  (p=0.008 n=5+5)
GDA/power-8          425ms ± 1%      906ms ± 1%    +113.30%  (p=0.008 n=5+5)
GDA/powersqrt-8      889ms ± 1%     2066ms ± 2%    +132.34%  (p=0.008 n=5+5)
GDA/randoms-8       8.57ms ± 0%  1354.78ms ± 1%  +15701.80%  (p=0.016 n=4+5)
GDA/squareroot-8    71.9ms ± 1%    213.9ms ± 1%    +197.48%  (p=0.008 n=5+5)
GDA/cuberoot-apd-8  4.18ms ± 0%     9.89ms ± 1%    +136.57%  (p=0.008 n=5+5)
```

It was mentioned in #13 that this algorithm is fast for large operands. I also performed a number of benchmarks with very large (>100 digits) operands in just the numerator, denominator, and both. In no case was I ever able to produce a benchmark that was faster using this method.

Some ideas about why this is: the first part reduces the denominator to [0.5, 1]. Normally this can be done very quickly by simply bit shifting as many times as needed if you are in base 2. However, being in base 10, we cannot do that, and must instead multiply the inputs each time. Ok I just did some benchmarks where the input is already within that range and it's still slower in the various cases I tested.

Another possibility is that this method has to use a precision 3 greater than that needed to have enough precision to round correctly. I really don't think it's this but who knows.

A third idea is the use of the loop convergence. Newton-Raphson is able to compute the number of iterations needed: maybe the loop is doing a few too many? My guess is this isn't a problem.

Anyone have other ideas or see things I'm doing clearly wrong? I used the psuedocode from wikipedia; https://en.wikipedia.org/wiki/Division_algorithm#Pseudocode.

@eisenstatdavid @RaduBerinde tagging you in case you're interested in this since ya'll are into algorithms.