@RaduBerinde 2017-09-14
I don't see the problem with the code, it follows the pseudocode AFAICT. I would add some printfs that show all iterations of the loop to see if we are doing too many iterations, and if the number of iterations is similar to what they say `(log2((P+1)/log2(17))` (note that `P` is binary precision though)

I would search for other adaptations of the algorithm that are tailored to base-10 decimal (the pseudocode as written seems to be done with base-2 floating-point in mind).

---

Review status: 0 of 2 files reviewed at latest revision, 2 unresolved discussions.

---

*[context.go, line 308 at r1](https://reviewable.io:443/reviews/cockroachdb/apd/72#-Ku-PosGmhSPqD_6TdGS:-Ku-PosGmhSPqD_6TdGT:bq8vlmq) ([raw file](https://github.com/cockroachdb/apd/blob/d948619898b5337d08258ec0468ef67df3df7d71/context.go#L308)):*
> ```Go
> 	D.Negative = false
> 	exp2 := 0
> 	for D.Cmp(decimalHalf) < 0 {
> ```

I would also check if these loops aren't taking a lot of time (we could look at multiplying/dividing by `2^something` first; or, we could calculate an initial exponent value from the base-10 exponent of D)

---

*[context.go, line 325 at r1](https://reviewable.io:443/reviews/cockroachdb/apd/72#-Ku-OYywZmDj6MDJcbJx:-Ku-OYywZmDj6MDJcbJy:b-9dg4mc) ([raw file](https://github.com/cockroachdb/apd/blob/d948619898b5337d08258ec0468ef67df3df7d71/context.go#L325)):*
> ```Go
> 	X0 := new(Decimal)
> 	for loop := nc.newLoop("quo", X, c.Precision, 1); ; {
> 		// X = X + X * (1 - D * X)
> ```

Why not `X = X * (2 - D * X)`? My guess is that the pseudocode we're following is written with floating-point in mind, it might make more sense in that context (?)

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/apd/72#-:-Ku-P9BSqAkd0GqfjQkn:bva2gzk)*
<!-- Sent from Reviewable.io -->