@mjibson 2017-09-15
Review status: 0 of 2 files reviewed at latest revision, 2 unresolved discussions.

---

*[context.go, line 308 at r1](https://reviewable.io:443/reviews/cockroachdb/apd/72#-Ku-PosGmhSPqD_6TdGS:-Ku3WSw1yeNQ4J4P0_0O:b-6nvfz7) ([raw file](https://github.com/cockroachdb/apd/blob/d948619898b5337d08258ec0468ef67df3df7d71/context.go#L308)):*
<details><summary><i>Previously, RaduBerinde wrote…</i></summary><blockquote>

I would also check if these loops aren't taking a lot of time (we could look at multiplying/dividing by `2^something` first; or, we could calculate an initial exponent value from the base-10 exponent of D)
</blockquote></details>

I did a test about this by benchmarking inputs that were already within [0.5, 1], and it was still slower. But I do think moving by a factor of 10 would be an improvement in the general case since it's so easy to modify the exponent.

---

*[context.go, line 325 at r1](https://reviewable.io:443/reviews/cockroachdb/apd/72#-Ku-OYywZmDj6MDJcbJx:-Ku3Wg-eTilnuHIN9DRZ:bskngl3) ([raw file](https://github.com/cockroachdb/apd/blob/d948619898b5337d08258ec0468ef67df3df7d71/context.go#L325)):*
The wikipedia article addresses this:

> From a computation point of view, the expressions {\displaystyle X_{i+1}=X_{i}+X_{i}(1-DX_{i})} X_{i+1}=X_{i}+X_{i}(1-DX_{i}) and {\displaystyle X_{i+1}=X_{i}(2-DX_{i})} X_{i+1}=X_{i}(2-DX_{i}) are not equivalent. To obtain a result with a precision of n bits while making use of the second expression, one must compute the product between {\displaystyle X_{i}} X_{i} and {\displaystyle (2-DX_{i})} (2-DX_{i}) with double the required precision (2n bits). In contrast, the product between {\displaystyle X_{i}} X_{i} and {\displaystyle (1-DX_{i})} (1-DX_{i}) need only be computed with a precision of n bits.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/apd/72)*
<!-- Sent from Reviewable.io -->