@mjibson 2017-09-15
As a test (https://github.com/cockroachdb/apd/commit/9af50e51015859ee380440bc1d136c4889c2cd6d) I implemented Goldschmidt division using the binomial simplification (https://en.wikipedia.org/wiki/Division_algorithm#Binomial_theorem). It performed slightly better than (about 7% faster) Newton-Raphson, but still real slow:

```
name \ time/op   old.txt     newton.txt   goldschmidt.txt
GDA/divide-4     887µs ±16%  6672µs ± 2%      6171µs ± 3%
GDA/divideint-4  120µs ±11%   140µs ±25%       141µs ±19%
```

It has one less operation per loop (a Sub), but the same number of multiplications. In order to pass the tests it has to bump the precision up pretty high (`c.Precision*2+1`). I don't have a lot of ideas about how to make this 10x faster to compete with what we have now, so unless someone else wants to take a stab, I'm going to close the issue in a few days.